Separate ResultData from Result
